app/requirements.txt
paho-mqtt>=1.6
pyserial
sds011lib
python-aqi
